automation/requirements.txt
# Python 패키지 목록
# pip install -r requirements.txt

# AI API
google-generativeai==0.3.0

# YouTube
youtube-transcript-api==0.6.0

# HTTP 요청
requests==2.31.0

# 스케줄러
schedule==1.2.0

# 유틸리티
python-dotenv==1.0.0
